refactor(material/list): fix action list CSS styles conflict with MDC-based list
The Angular Material action list has been styled using an element selector,
instead of using the corresponding CSS class. This meant that the MDC-based
list seemed to have a working action list, but in reality it was just working due
to the leaked styles from the non-MDC list.

This fixes the selector in the Angular Material list, and adds the necessary styles
for the action list to the MDC list styles.