build: fix tarball generation for unit tests
```
➤ YN0001: │ Error: EEXIST: file already exists, mkdir 'node_modules/@angular/material/testing'
    at makeError ([worker eval]:474:24)
    at EEXIST ([worker eval]:498:10)
    at ZipFS.mkdirSync ([worker eval]:2378:13)
    at extractArchiveTo ([worker eval]:15405:20)
    at async MessagePort.<anonymous> ([worker eval]:15451:3)
```

This is an error I received from Yarn when I passed it a file:<bazel-built-tar>.
The tar was built from the npm_package using pkg_tar and strip_prefix = "npm_package"
(to keep the tar flat obviously). Interestingly the folder mentioned in the error is
not even part of the NPM package (just internal for dev). It looks like pkg_tar included
some metadata of directories/context into the tarball, even if it only includes files
for <cwd>/npm_package (due to the strip prefix)

The fix is to use package_dir instead since that means that only files inside the
npm_package directory are visible to the tar creation. An interesting finding!